Module ggp_rs::player

Ready to use GGP players

Structs

AlphaBetaPlayer
An alpha beta search player with a bounded depth. This player should only be used for 2 player games.

CompDelibPlayer
A bounded compulsive deliberation player. This should only be used for single player games

LegalPlayer
A player that returns the first legal move it finds

McsPlayer
A Monte Carlo search player. This player should only be used for 2 player, constant sum, turn based games.

MinimaxPlayer
A bounded minimax player. This player should only be used for 2 player, constant sum, turn based games.

RandomPlayer
A player that chooses a random move from all possible of legal moves