Struct ggp_rs::player::MinimaxPlayer

pub struct MinimaxPlayer {
    // some fields omitted
}

A bounded minimax player. This player should only be used for 2 player, constant sum, turn based games.

Methods

impl MinimaxPlayer

fn new() -> MinimaxPlayer

Returns a minimax player with the default bounds of 0 and 100

fn with_bounds(min_bound: u8, max_bound: u8) -> MinimaxPlayer

Returns a minimax player with the given min and max bounds